requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.6.1
psycopg[binary]==3.1.18
psycopg_pool==3.2.1
